requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
aiosqlite==0.20.0
alembic==1.8.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.7.0
apprise==1.9.1
asgi-lifespan==2.1.0
asttokens==3.0.0
asyncpg==0.30.0
attrs==24.3.0
bcrypt==4.0.1
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
coolname==2.2.0
croniter==5.0.1
cryptography==44.0.0
cycler==0.12.1
databricks-sdk==0.38.0
dateparser==1.2.0
debugpy==1.8.9
decorator==5.1.1
Deprecated==1.2.15
distro==1.9.0
docker==7.1.0
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.6
fastjsonschema==2.21.1
Flask==3.1.0
fonttools==4.55.2
frozenlist==1.5.0
fsspec==2024.12.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.36.0
graphene==3.4.3
graphql-core==3.2.5
graphql-relay==3.2.0
graphviz==0.20.3
greenlet==3.1.1
griffe==1.5.1
gunicorn==23.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
humanize==4.11.0
hyperframe==6.0.1
idna==3.10
imbalanced-learn==0.13.0
imblearn==0.0
importlib_metadata==8.5.0
ipinfo==5.1.1
ipykernel==6.29.5
ipython==8.30.0
ipywidgets==8.1.5
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jinja2-humanize-extension==0.4.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
Mako==1.3.7
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mlflow==2.18.0
mlflow-skinny==2.18.0
multidict==6.1.0
narwhals==1.19.1
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==2.1.3
nvidia-nccl-cu12==2.23.4
oauthlib==3.2.2
opentelemetry-api==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-semantic-conventions==0.49b2
orjson==3.10.12
packaging==24.2
pandas==2.2.3
parso==0.8.4
passlib==1.7.4
pathspec==0.12.1
pendulum==3.0.0
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
prefect==3.1.10
prefect-docker==0.6.2
prometheus_client==0.21.1
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==5.29.1
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.8.2
pydantic-extra-types==2.10.1
pydantic-settings==2.6.1
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.7.0
PyMySQL==1.1.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
python-slugify==8.0.4
python-socks==2.6.0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
readchar==4.2.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
secure==0.3.0
setuptools==75.6.0
shellingham==1.5.4
six==1.17.0
sklearn-compat==0.1.3
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.36
SQLAlchemy-Utils==0.41.2
sqlmodel==0.0.18
sqlparse==0.5.2
stack-data==0.6.3
starlette==0.41.3
streamlit==1.41.1
tenacity==9.0.0
text-unidecode==1.3
threadpoolctl==3.5.0
time-machine==2.16.0
toml==0.10.2
tornado==6.4.2
traitlets==5.14.3
typer==0.13.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
ucimlrepo==0.0.7
ujson==5.10.0
urllib3==2.2.3
uvicorn==0.34.0
uvloop==0.21.0
watchdog==6.0.0
watchfiles==1.0.3
wcwidth==0.2.13
websockets==13.1
Werkzeug==3.1.3
widgetsnbextension==4.0.13
wrapt==1.17.0
xgboost==2.1.3
yarl==1.18.3
zenml==0.71.0
zipp==3.21.0
